taskmanagerproject/requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
